### 2025-12-06 #9559 @editorialbot
[PRE REVIEW]: IAMAP: Unlocking Deep Learning in QGIS for non-coders and limited computing resources
**Submitting author:** @ptresson (<a href="http://orcid.org/0000-0002-1275-4673">Paul Tresson</a>)
**Repository:** https://github.com/umr-amap/iamap
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.7.0
