### 2025-12-16 b5f28a0
cleanup refs and dois
