+++ paper.bib
@@ -14,6 +14,7 @@
   number={4},
   pages={8--36},
   year={2017},
+  doi={10.1109/mgrs.2017.2762307},
   publisher={IEEE}
 }
 
@@ -34,6 +35,7 @@
   number={1},
   pages={157--193},
   year={2023},
+  doi={10.1080/01431161.2022.2161856},
   publisher={Taylor \& Francis}
 }
 
@@ -44,6 +46,7 @@
   volume={125},
   pages={103569},
   year={2023},
+  doi={10.31223/x52h3b},
   publisher={Elsevier}
 }
 
@@ -62,283 +65,8 @@
   number={3},
   pages={252},
   year={2024},
+  doi={10.3390/e26030252},
   publisher={MDPI}
-}
-
-@ARTICLE{Fortunato2010,
-  author  = {Fortunato, S.},
-  title   = {Community detection in graphs},
-  journal = {Phys. Rep.-Rev. Sec. Phys. Lett.}, 
-  volume  = {486},
-  year    = {2010},
-  pages   = {75-174}
-}
-
-@ARTICLE{NewmanGirvan2004,
-  author  = {Newman, M. E. J. and Girvan, M.},
-  title   = {Finding and evaluating community structure in networks},
-  journal = {Phys. Rev. E.}, 
-  volume  = {69},
-  year    = {2004},
-  pages   = {026113}
-}
-
-@ARTICLE{Vehlowetal2013,
-  author  = {Vehlow, C. and Reinhardt, T. and Weiskopf, D.},
-  title   = {Visualizing Fuzzy Overlapping Communities in Networks},
-  journal = {IEEE Trans. Vis. Comput. Graph.}, 
-  volume  = {19},
-  year    = {2013},
-  pages   = {2486-2495}
-}
-
-@ARTICLE{Raghavanetal2007,
-  author  = {Raghavan, U. and Albert, R. and Kumara, S.},
-  title   = {Near linear time algorithm to detect community structures in large-scale networks},
-  journal = {Phys. Rev E.}, 
-  volume  = {76},
-  year    = {2007},
-  pages   = {036106}
-}
-
-@ARTICLE{SubeljBajec2011a,
-  author  = {\v{S}ubelj, L. and Bajec, M.},
-  title   = {Robust network community detection using balanced propagation},
-  journal = {Eur. Phys. J. B.}, 
-  volume  = {81},
-  year    = {2011},
-  pages   = {353-362}
-}
-
-@ARTICLE{Louetal2013,
-  author  = {Lou, H. and Li, S. and Zhao, Y.},
-  title   = {Detecting community structure using label propagation with weighted coherent neighborhood propinquity},
-  journal = {Physica A.}, 
-  volume  = {392},
-  year    = {2013},
-  pages   = {3095-3105}
-}
-
-@ARTICLE{Clausetetal2004,
-  author  = {Clauset, A. and Newman, M. E. J. and Moore, C.},
-  title   = {Finding community structure in very large networks},
-  journal = {Phys. Rev. E.}, 
-  volume  = {70},
-  year    = {2004},
-  pages   = {066111}
-}
-
-@ARTICLE{Blondeletal2008,
-  author  = {Blondel, V. D. and Guillaume, J. L. and Lambiotte, R. and Lefebvre, E.},
-  title   = {Fast unfolding of communities in large networks},
-  journal = {J. Stat. Mech.-Theory Exp.}, 
-  volume  = {2008},
-  year    = {2008},
-  pages   = {P10008}
-}
-
-@ARTICLE{SobolevskyCampari2014,
-  author  = {Sobolevsky, S. and Campari, R.},
[…]
