### 2026-01-16 #9559 @crvernon
Also please ensure that the word count of your paper is within JOSS requirements.  Thank you.
