### 2026-01-19 #9559 @ptresson
> [@editorialbot](https://github.com/editorialbot) invite [@cheginit](https://github.com/cheginit) as editor
>
> 👋 [@cheginit](https://github.com/cheginit) are you available to take this one on as editor?
>
> [@ptresson](https://github.com/ptresson) will need to adjust the submission to comply with the new JOSS standards, but you should be able to get started directly after if you are available.

Hi, I've noted the new standards and word count. I'll review the paper to ensure compliance shortly (I'm hopping beginning of the week).
