+++ paper.md
@@ -32,6 +32,77 @@ bibliography: paper.bib
 
 ---
 
+# Summary 
+
+# Statement of need
+
+The integration of remote sensing data with deep learning approaches is
+currently revolutionizing Earth observation sciences, leading to
+significant qualitative and quantitative improvements in large-scale
+predictions [@zhu2017deep; @yuan2020deep; @yasir2023coupling]. However,
+this revolution comes with a number of challenges. First, over the past
+decade, most deep learning applications have been highly data-demanding,
+requiring extensive manual labeling with typically more than one hundred
+thousands labeled points [@safonova2023ten]. In most ecological and
+environmental science studies, constructing such a large reference
+dataset, through *e.g.*, ground observations or photo-interpretation,
+remains a major barrier to the implementation of deep learning
+approaches.
+
+Recent developments in deep learning state of the art has seen an evolution
+towards the use of large foundation models, trained on large scale datasets 
+in an unsupervised way and capable of very good few-shot performances (*i.e.*
+without expensive supervised training of the network).
+[@ericsson2021well]. The main difference between a pre-trained
+self-supervised learning (SSL) model and a pre-trained supervised model
+lies in their training objectives: SSL models are not constrained by
+predefined labels and are therefore free to explore and encode the
+intrinsic structure and diversity of the data, often resulting in more
+general and transferable representations. In contrast, supervised models
+are explicitly optimized to perform a specific user-defined task, which
+can lead to highly specialized representations that may overlook other
+meaningful features in the data. As such, SSL foundation models can
+perform well even in low-shot or zero-shot tasks, *i.e.* using the model
+as is, with few or no training data. Consequently, SSL models are
+considered particularly promising for remote sensing tasks, as
+demonstrated by recent works and initiatives
+[@jakubik2023prithvi; @cong2023satmae; @xiong2024neural; @marsocci2024pangaea].
+
+
+# State of the field
+
+With the democratization of deep learning, some developers have already
+worked on the integration of deep learning models in geographic
+information systems such as the open-source and widely used QGIS
+software [@QGIS]. However, at the time of writing, these solutions
+mostly focus on fine-tuning models or using a model in inference only
+[*e.g.* see @aszkowski2023deepness; @zhao2023geosam]. Then, they are
+only usable by users with access to high-end computing power, extensive
+dataset, on interested in a task for which a specific model was already
+trained.
+
+# Software design
+
+The plugin aims at allowing two main tasks: 
+1) feeding a tiled raster through a pre-trained deep learning model,
+    typically via the *timm* and *huggingface* library.
+2) manipulating the produced features with common machine learning tools,
+    provided by the *scikit-learn* library.
+
+these tasks are separated into five modules depending on the type of models
+and operations the user wants to perform: deep learning feature extraction,
+dimension reduction, clustering, similarity computation, or fitting a machine
+learning model in a supervised way (see.Fig. \autoref{fig:functionalities}).
+        
+The 
+
+# Research impact statement
+
+# AI usage disclosure
+
+
+==================================================
+
 # Summary
 
   **1.** Remote sensing has entered a new era with the rapid development
