+++ paper.md
@@ -94,470 +94,469 @@ and operations the user wants to perform: deep learning feature extraction,
 dimension reduction, clustering, similarity computation, or fitting a machine
 learning model in a supervised way (see.Fig. \autoref{fig:functionalities}).
         
-The 
 
 # Research impact statement
 
 # AI usage disclosure
 
 
-==================================================
-
-# Summary
-
-  **1.** Remote sensing has entered a new era with the rapid development
-  of artificial intelligence approaches. However, the implementation of
-  deep learning has largely remained restricted to specialists and has
-  been impractical because it often requires (i) large reference
-  datasets for model training and validation; (ii) substantial computing
-  resources; and (iii) strong coding skills.
-
-  **2.** Here, we introduce IAMAP, a user-friendly QGIS plugin that
-  addresses these three challenges in an easy yet flexible way. IAMAP
-  builds on recent advancements in self-supervised learning strategies,
-  which now provide robust feature extractors, often referred to as
-  foundation models. These generalist models can often be reliably used
-  in few-shot or zero-shot scenarios (*i.e.*, with little to no
-  fine-tuning).
-
-  **3.** IAMAP's interface allows users to streamline several key steps
-  in remote sensing image analysis: (i) extracting image features using
-  a wide range of deep learning architectures; (ii) reducing
-  dimensionality with built-in algorithms; (iii) performing clustering
-  on features or their reduced representations; (iv) generating feature
-  similarity maps; and (v) calibrating and validating supervised machine
-  learning models for prediction.
-
-  **4.** By enabling non-AI specialists to leverage the high-quality
-  features provided by recent deep learning approaches without requiring
-  GPU capacity or extensive reference datasets, IAMAP contributes to the
-  democratization of computationally efficient and energy-conscious deep
-  learning methods.
-
-  **Keywords:** Remote sensing, Self-supervised learning, Foundation
-  models, Machine learning, Artificial Intelligence, Consumer hardware.
-
-# Statement of need
-
-The integration of remote sensing data with deep learning approaches is
-currently revolutionizing Earth observation sciences, leading to
-significant qualitative and quantitative improvements in large-scale
-predictions [@zhu2017deep; @yuan2020deep; @yasir2023coupling]. However,
-this revolution comes with a number of challenges. First, over the past
-decade, most deep learning applications have been highly data-demanding,
-requiring extensive manual labeling with typically more than one hundred
-thousands labeled points [@safonova2023ten]. In most ecological and
-environmental science studies, constructing such a large reference
-dataset, through *e.g.*, ground observations or photo-interpretation,
-remains a major barrier to the implementation of deep learning
-approaches. Second, a common obstacle to the adoption of deep learning
-is the computing power required to train a model. Training a deep
-learning model is indeed highly resource-intensive, primarily due to the
-backpropagation step [see @goodfellow2016deep]. As a result, modern deep
-learning architectures are virtually impossible to train without
-substantial local Graphics Processing Unit (GPU) capacity or access to
-high-end computing clusters. Last but not least, implementing deep
-learning approaches typically requires at least basic coding skills,
-which has so far restricted their use to users with a minimal background
-in computer science.
-
-The recent development of self-supervised learning (SSL) approaches is a
-game-changer in the deep learning domain, as exemplified by the success
-of models like BERT and ChatGPT in natural language processing
-[@devlin2018bert; @achiam2023gpt]. In SSL, the model starts by learning
-features describing a dataset via a pretext task that does not require a
-label. In computer vision, several SSL strategies have been proposed,
-typically belonging to two main categories: contrastive or generative
-learning. In contrastive learning, several networks view transformed
[…]
