### 2026-02-24 9344790
software design section, develop research impact, add back references etc..
