+++ paper.md
@@ -90,480 +90,79 @@ The plugin aims at allowing two main tasks:
     provided by the *scikit-learn* library.
 
 these tasks are separated into five modules depending on the type of models
-and operations the user wants to perform: deep learning feature extraction,
-dimension reduction, clustering, similarity computation, or fitting a machine
-learning model in a supervised way (see.Fig. \autoref{fig:functionalities}).
-        
+and operations the user wants to perform: deep learning feature extraction(see.
+Fig. \autoref{fig:backbones}), dimension reduction, clustering, similarity 
+computation, or fitting a machine learning model in a supervised way (see.Fig.
+ \autoref{fig:functionalities}).
+
+![The five main modules of the IAMAP
+plugin. \label{fig:functionalities}](figures/plugin.png)
+
+![A sentinel 2 image of a forested landscape in Thailand (Khao
+Banthat Wildlife Sanctuary; Lat 7.53°, Lon 99.82°) processed by
+different backbones. The top row represents the first three feature
+dimensions output by the models (which may not be the most informative).
+The second row shows a 3D PCA of the features mapped to the red, green
+and blue channel respectively. The third row shows a projection using a
+3D T-SNE.\label{fig:backbones}](figures/backbones.png)
+
+- **The deep learning feature extraction module** allows to feed a raster into 
+a ViT-like backbone. The user can chose a raster loaded in QGIS, a pre-trained 
+model that is downloaded from timm/huggingface or pass local pre-trained 
+weights. The raster is then tiled according to user defined rules and fed into 
+the model using a light fork of the *torchgeo library*. Resulting features are 
+saved into a geotiff file and can then be used for further analysis.
+- **The dimension reduction module** uses *scikit-learn* API to feed a raster 
+into a model and collect the resulting raster. All algorithms available in 
+the *scikit-learn* `decomposition` and `cluster` modules that have common APIs 
+(namely, a `fit()`, a `transform()`, or a `fit_transform()` method) are 
+available. The parameters can be fed following a json format.
+- **The clustering module** works in a similar manner with algorithms 
+available in the *scikit-learn* `cluster` module sharing common APIs (namely, 
+a `fit()`, a `predict()`, or a `fit_predict()` method).
+- **The similarity module** computes the similarity between a raster and user-
+defined template pixels (via a vector layer).
+- **The machine learning module** allows the user to fit a supervised machine 
+learning model available in the *sciki-learn* `ensemble` and `neighbors` 
+modules. Hyper-parameters (*e.g.* validation scheme, labelled data… can be 
+defined by the user).
+
+A core constraint of the development is for the plugin to be accessible for a 
+non-coding user without a GPU and with limited internet access. As such, all 
+design and development is done a CPU only laptop to assess usability with no 
+GPU. Various optimizations are available, such as quantization, scheduled 
+pauses and progress save on disk. We used code from the 
+[deepness plugin](https://github.com/PUTvision/qgis-plugin-deepness/) to 
+provide automated dependencies installation dialog.
 
 # Research impact statement
 
-The plugin is used internally and has been presented in several webinars. It 
-has also be used for research to be published soon by Malkassian et. al. (A 
-preview of the results [is available in the documentation](https://iamap.readthedocs.io/en/latest/examples.html#mapping-land-cover-change-in-la-reunion-island)).
+The plugin is used internally and has been presented in several webinars with 
+strong positive feedback from potential users. It has also be used for research
+to be published soon by Malkassian et. al. (A preview of the results 
[…]
