### 2026-02-26 #9559 @ptresson
Hi,
just to inform you that I've updated the article with the new JOSS format. I'm currently waiting on feedback from my co-authors before merging this new version.
