### 2026-03-02 #9559 @nkrusch
👋 Hello @hydrogeohc, @florianfranz, @jingxuehan0707, @goergen95, and @YOUNKAP

would you be interested in reviewing this submission for JOSS?
