### 2026-03-13 262bdc3
add summary, integrate comments from MRM
