### 2026-03-13 #9559 @ptresson
Hi,

we will do a final review of the paper with the new format this evening/Monday morning. I hope to merge changes on the `joss` branch by Monday.
