### 2026-03-16 #9559 @nkrusch
Hi @Digdgeo, all the reviewer slots are already allocated, unfortunately. But I will keep you in mind, and notify you again next time I have some GIS-related submission.
