### 2026-03-30 #10329 @nkrusch
:wave:  @florianfranz, @AMSANJEEV28, @jhwohlgemuth - Thank you for agreeing to review this submission!

As a reviewer, your first step is to **create a checklist** by entering

as a new comment in this thread. The checklist contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied.

The [JOSS review](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#10329` so that a link is created to this thread. Feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We ask reviewers to complete their reviews in 4-6 weeks, and we generally ask the same commitment from authors: they should respond to your feedback within 2 weeks and complete requested changes within 4-6 weeks, unless otherwise negotiated with the editor. Please let me know if you need more time. You may also use [EditorialBot](https://joss.readthedocs.io/en/latest/editorial_bot.html) (our bot) to set automatic reminders if you know you'll be away for a known period of time.
