### 2026-04-22 #10329 @AMSANJEEV28
Overall software meets most JOSS requirements and is well structured .
However, a few areas need improvement:
1. Example usage is limited.
2. Community guidelines,
3. References are incomplete or missing DOIs.
@nkrusch
