### 2026-04-22 #10329 @nkrusch
@ptresson, please fix:

> MISSING DOIs: 10.52202/068431-0015 may be a valid DOI for title: SatMAE: Pre-training Transformers for Temporal and...
