### 2026-04-23 d1b0ee0
update SatMAE ref with neurips rather than arxiv
neurips procedings are publshed as chapters and have no DOI cf. https://academia.stackexchange.com/questions/151409/why-are-there-no-dois-for-neurips-papers
