+++ paper.bib
@@ -218,13 +218,13 @@
     year            = {2023}
 }
 
-@misc{cong2023satmae,
-      title={SatMAE: Pre-training Transformers for Temporal and Multi-Spectral Satellite Imagery}, 
-      author={Yezhen Cong and Samar Khanna and Chenlin Meng and Patrick Liu and Erik Rozi and Yutong He and Marshall Burke and David B. Lobell and Stefano Ermon},
-      year={2023},
-      eprint={2207.08051},
-      archivePrefix={arXiv},
-      primaryClass={id='cs.CV' full_name='Computer Vision and Pattern Recognition' is_active=True alt_name=None in_archive='cs' is_general=False description='Covers image processing, computer vision, pattern recognition, and scene understanding. Roughly includes material in ACM Subject Classes I.2.10, I.4, and I.5.'}
+@article{cong2022satmae,
+  title={Satmae: Pre-training transformers for temporal and multi-spectral satellite imagery},
+  author={Cong, Yezhen and Khanna, Samar and Meng, Chenlin and Liu, Patrick and Rozi, Erik and He, Yutong and Burke, Marshall and Lobell, David and Ermon, Stefano},
+  journal={Advances in Neural Information Processing Systems},
+  volume={35},
+  pages={197--211},
+  year={2022}
 }
 
 @inproceedings{braham2022self,
