### 2026-04-23 #10329 @ptresson
> [@ptresson](https://github.com/ptresson), please fix:
>
> > MISSING DOIs: 10.52202/068431-0015 may be a valid DOI for title: SatMAE: Pre-training Transformers for Temporal and...

Hi,
this should be done with d1b0ee02ebce8fe8d8253b937d3229cdf4ddfdff on the original repo.
