### 2026-04-23 2eaf173
add doi proposed by editorial bot for satmae
