+++ paper.bib
@@ -223,6 +223,7 @@
   author={Cong, Yezhen and Khanna, Samar and Meng, Chenlin and Liu, Patrick and Rozi, Erik and He, Yutong and Burke, Marshall and Lobell, David and Ermon, Stefano},
   journal={Advances in Neural Information Processing Systems},
   volume={35},
+  doi={10.52202/068431-0015},
   pages={197--211},
   year={2022}
 }
