### 2026-04-23 #10329 @ptresson
> ```
> ❌ MISSING DOIs
>
> - 10.52202/068431-0015 may be a valid DOI for title: Satmae: Pre-training transformers for temporal and...
> ```

I've added the DOI proposed by the bot as it leads to the same version of the paper published in neurips
