### 2026-04-29 #10329 @florianfranz
This submission meets JOSS requirements. Very happy with how this plugin is designed and how the documentation is written. The paper summarizes very well the challenges, the need and how this plugin addresses these needs.

I agree with comments from @AMSANJEEV28 in https://github.com/openjournals/joss-reviews/issues/10329#issuecomment-4294716881 about the need to include the possibility to redo the steps outlined in the videos for the tools, as we currently cannot test if the software works or not. Also noticed a missing citation resulting in a question mark on line 47 (at least in the latest version of the manuscript generated with the editorialboat in https://github.com/openjournals/joss-reviews/issues/10329#issuecomment-4324851146.

For these reasons, I have left some unticked boxes in my review that pertain to (i) data sharing (ii) software performances and, (iii) bibliographic references. This should be reflected by minor edits to the docs if instructions on where to get the data are added allowing to reuse the software from end-to-end.
