### 2026-04-29 #10329 @nkrusch
Thank you for your feedback @florianfranz and @AMSANJEEV28! The JOSS review process is to iterate over these deficiencies until all review criteria is satisfactory. So, @ptresson, let us know when an update is available.
