### 2026-05-20 06b80b2
add link to example raster file and description of the encoding process in the docs
