### 2026-05-25 #10329 @AMSANJEEV28
Thank you for the updates.
I reviewed the revised documentation and paper. The added dataset access, workflow instructions, and corrected references address my earlier concerns regarding example usage and citations.

The submission now satisfies the JOSS review criteria from my side.
