### 2026-06-01 d704e7b
link to template shapefile in the docs
