### 2026-06-03 65ad12b
update paper with comments from nkrusch
