+++ paper.bib
@@ -3,7 +3,7 @@
     author = {{QGIS Development Team}},
     organization = {Open Source Geospatial Foundation},
     year = {2025},
-    url = {http://qgis.org},
+    url = {https://qgis.org},
   }
 
 @article{zhu2017deep,
@@ -39,16 +39,6 @@
   publisher={Taylor \& Francis}
 }
 
-@article{safonova2023ten,
-  title={Ten deep learning techniques to address small data problems with remote sensing},
-  author={Safonova, Anastasiia and Ghazaryan, Gohar and Stiller, Stefan and Main-Knorn, Magdalena and Nendel, Claas and Ryo, Masahiro},
-  journal={International Journal of Applied Earth Observation and Geoinformation},
-  volume={125},
-  pages={103569},
-  year={2023},
-  doi={10.31223/x52h3b},
-  publisher={Elsevier}
-}
 
 @misc{goodfellow2016deep,
   title={Deep learning},
@@ -281,14 +271,16 @@
     month           = oct,
     title           = {{Foundation Models for Generalist Geospatial Artificial Intelligence}},
     journal         = {Preprint Available on arxiv:2310.18660},
-    year            = {2023}
+    year            = {2023},
+    url             = {https://arxiv.org/pdf/2310.18660},
 }
 
 @article{xiong2024neural,
   title={Neural Plasticity-Inspired Foundation Model for Observing the {Earth} Crossing Modalities},
   author={Xiong, Zhitong and Wang, Yi and Zhang, Fahong and Stewart, Adam J and Hanna, Jo{\"e}lle and Borth, Damian and Papoutsis, Ioannis and Saux, Bertrand Le and Camps-Valls, Gustau and Zhu, Xiao Xiang},
   journal={arXiv preprint arXiv:2403.15356},
-  year={2024}
+  year={2024},
+  url={https://arxiv.org/pdf/2403.15356?},
 }
 
 @article{vaswani2017attention,
@@ -467,4 +459,38 @@ keywords = {QGIS, Deep learning, Remote sensing, Segmentation, Object detection}
   author={Wang, Yi and Braham, Nassim Ait Ali and Xiong, Zhitong and Liu, Chenying and Albrecht, Conrad M and Zhu, Xiao Xiang},
   journal={arXiv preprint arXiv:2211.07044},
   year={2022}
+}
+
+@article{sklearn,
+  author  = {Fabian Pedregosa and Ga{{\"e}}l Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and {{\'E}}douard Duchesnay},
+  title   = {Scikit-learn: Machine Learning in Python},
+  journal = {Journal of Machine Learning Research},
+  year    = {2011},
+  volume  = {12},
+  number  = {85},
+  pages   = {2825--2830},
+  url     = {http://jmlr.org/papers/v12/pedregosa11a.html}
+}
+
+@article{stewart2024torchgeo,
+    author = {Stewart, Adam J. and Robinson, Caleb and Corley, Isaac A. and Ortiz, Anthony and Lavista Ferres, Juan M. and Banerjee, Arindam},
+    title = {{TorchGeo}: Deep Learning With Geospatial Data},
+    journal = {ACM Transactions on Spatial Algorithms and Systems},
+    volume = {11},
+    number = {4},
+    pages = {1--28},
+    month = aug,
+    year = {2025}
+}
+
+@article{safonova2023ten,
[…]
