### 2026-06-03 a2ac65a
small typo fixes on citation and add url for timm and hugginface citations
