+++ paper.bib
@@ -149,6 +149,7 @@
   title={An image is worth 16x16 words: Transformers for image recognition at scale},
   author={Dosovitskiy, Alexey and Beyer, Lucas and Kolesnikov, Alexander and Weissenborn, Dirk and Zhai, Xiaohua and Unterthiner, Thomas and Dehghani, Mostafa and Minderer, Matthias and Heigold, Georg and Gelly, Sylvain and others},
   journal={arXiv preprint arXiv:2010.11929},
+  url={https://arxiv.org/pdf/2010.11929/100},
   year={2020}
 }
 
@@ -304,6 +305,7 @@
   title={Huggingface's transformers: State-of-the-art natural language processing},
   author={Wolf, T},
   journal={arXiv preprint arXiv:1910.03771},
+  url = {https://arxiv.org/pdf/1910.03771},
   year={2019}
 }
 
