+++ paper.bib
@@ -492,7 +492,7 @@ volume = {125},
 pages = {103569},
 year = {2023},
 issn = {1569-8432},
-doi = {https://doi.org/10.1016/j.jag.2023.103569},
+doi = {10.1016/j.jag.2023.103569},
 url = {https://www.sciencedirect.com/science/article/pii/S156984322300393X},
 author = {Anastasiia Safonova and Gohar Ghazaryan and Stefan Stiller and Magdalena Main-Knorn and Claas Nendel and Masahiro Ryo},
 }
