+++ paper.md
@@ -35,10 +35,10 @@ bibliography: paper.bib
 
 # Summary 
 
-We introduce IAMAP, a user-friendly plugin  for the geographic information system QGIS that allows to 
+We introduce IAMAP, a user-friendly plugin for the geographic information system QGIS that allows users to 
 (1) tile raster images and feed them into pre-trained deep learning model to 
 extract image features; 
-(2) reduce features dimensionality; 
+(2) reduce feature dimensionality; 
 (3) perform clustering on features or their reduced representations; 
 (4) generate feature similarity maps; 
 (5) calibrate and validate supervised machine learning models to create maps.
@@ -47,8 +47,8 @@ by foundation models without requiring GPU capacity or extensive reference
 datasets, IAMAP contributes to the democratization of these computationally 
 efficient and energy-conscious methods.
 Development of the plugin is 
-[open sourced on GitHub](https://github.com/umr-amap/iamap). Documentation and 
-tutorials are available [on readthedocs](https://iamap.readthedocs.io/). 
+[open-source on GitHub](https://github.com/umr-amap/iamap). Documentation and 
+tutorials are available [on Read the Docs](https://iamap.readthedocs.io/). 
 
 # Statement of need
 
@@ -61,7 +61,7 @@ this revolution comes with a number of challenges.
 First, over the past
 decade, most deep learning applications have been highly data-demanding,
 requiring extensive manual labeling with typically more than one hundred
-thousands labeled points [@safonova2023ten]. In most ecological and
+thousand labeled points [@safonova2023ten]. In most ecological and
 environmental science studies, constructing such a large reference
 dataset, through *e.g.*, ground observations or photo-interpretation,
 remains a major barrier to the implementation of deep learning
@@ -69,7 +69,7 @@ approaches.
 
 Second, most deep learning developments have been conducted by and for 
 actors able to leverage significant computing power with the training of a 
-model being GPU-dependant. This creates a significant barrier to users who 
+model being GPU-dependent. This creates a significant barrier to users who 
 could otherwise benefit from inference-only deep learning applications.
 
 Finally, the use of state of the art models is still mostly confined to users 
@@ -80,7 +80,7 @@ ease the implementation of models.
 
 Recently, deep learning state of the art has seen an evolution
 towards the use of large foundation models, trained on large scale datasets 
-in an unsupervised way, and capable of very good few-shot performances (*i.e.*,
+in an unsupervised way, and capable of very good few-shot performance (*i.e.*,
 without expensive supervised training of the network)
 [@ericsson2021well]. The main difference between a pre-trained
 self-supervised learning (SSL) model and a pre-trained supervised model
@@ -104,10 +104,10 @@ software [@QGIS]. However, at the time of writing, these solutions
 mostly focus on fine-tuning models or using a model in inference only
 [*e.g.*, see @aszkowski2023deepness; @zhao2023geosam]. Then, they are
 only usable by users with access to high-end computing power, extensive
-dataset, on interested in a task for which a specific model was already
+dataset, or interested in a task for which a specific model was already
 trained.
 
-We have therefore developped a plugin that allows to use a variety of deep 
+We have therefore developed a plugin that allows users to use a variety of deep 
 learning models and manipulate their features in a user-friendly fashion.
 
 # Software design
@@ -115,10 +115,10 @@ learning models and manipulate their features in a user-friendly fashion.
 This plugin aims at allowing two main tasks:
 
 1. feeding a tiled raster through a pre-trained deep learning model,
-    typically via the *timm* and *huggingface* library [@rw2019timm; @wolf2019huggingface].
[…]
