### 2026-06-25 570b7dd
update refs according to joss's editor remarks
