+++ paper.bib
@@ -20,11 +20,12 @@
 
 @article{yuan2020deep,
   title={Deep learning in environmental remote sensing: Achievements and challenges},
-  author={Yuan, Qiangqiang and Shen, Huanfeng and Li, Tongwen and Li, Zhiwei and Li, Shuwen and Jiang, Yun and Xu, Hongzhang and Tan, Weiwei and Yang, Qianqian and Wang, Jiwen and others},
-  journal={Remote sensing of Environment},
+  author={Qiangqiang Yuan and Huanfeng Shen and Tongwen Li and Zhiwei Li and Shuwen Li and Yun Jiang and Hongzhang Xu and Weiwei Tan and Qianqian Yang and Jiwen Wang and Jianhao Gao and Liangpei Zhang},
+  journal={Remote Sensing of Environment},
   volume={241},
   pages={111716},
   year={2020},
+  doi={10.1016/j.rse.2020.111716},
   publisher={Elsevier}
 }
 @article{yasir2023coupling,
@@ -210,7 +211,7 @@
 }
 
 @article{cong2022satmae,
-  title={Satmae: Pre-training transformers for temporal and multi-spectral satellite imagery},
+  title={Sat{MAE}: Pre-training transformers for temporal and multi-spectral satellite imagery},
   author={Cong, Yezhen and Khanna, Samar and Meng, Chenlin and Liu, Patrick and Rozi, Erik and He, Yutong and Burke, Marshall and Lobell, David and Ermon, Stefano},
   journal={Advances in Neural Information Processing Systems},
   volume={35},
@@ -271,17 +272,17 @@
     author          = {Jakubik, Johannes and Roy, Sujit and Phillips, C. E. and Fraccaro, Paolo and Godwin, Denys and Zadrozny, Bianca and Szwarcman, Daniela and Gomes, Carlos and Nyirjesy, Gabby and Edwards, Blair and Kimura, Daiki and Simumba, Naomi and Chu, Linsong and Mukkavilli, S. Karthik and Lambhate, Devyani and Das, Kamal and Bangalore, Ranjini and Oliveira, Dario and Muszynski, Michal and Ankur, Kumar and Ramasubramanian, Muthukumaran and Gurung, Iksha and Khallaghi, Sam and Li, Hanxi (Steve) and Cecil, Michael and Ahmadi, Maryam and Kordi, Fatemeh and Alemohammad, Hamed and Maskey, Manil and Ganti, Raghu and Weldemariam, Kommy and Ramachandran, Rahul},
     month           = oct,
     title           = {{Foundation Models for Generalist Geospatial Artificial Intelligence}},
-    journal         = {Preprint Available on arxiv:2310.18660},
+    journal         = {Preprint Available on {a}r{X}iv:2310.18660},
     year            = {2023},
     url             = {https://arxiv.org/pdf/2310.18660},
 }
 
 @article{xiong2024neural,
-  title={Neural Plasticity-Inspired Foundation Model for Observing the {Earth} Crossing Modalities},
+  title={Neural Plasticity-Inspired Multimodal Foundation Model Earth Observation},
   author={Xiong, Zhitong and Wang, Yi and Zhang, Fahong and Stewart, Adam J and Hanna, Jo{\"e}lle and Borth, Damian and Papoutsis, Ioannis and Saux, Bertrand Le and Camps-Valls, Gustau and Zhu, Xiao Xiang},
   journal={arXiv preprint arXiv:2403.15356},
   year={2024},
-  url={https://arxiv.org/pdf/2403.15356?},
+  url={https://arxiv.org/pdf/2403.15356},
 }
 
 @article{vaswani2017attention,
@@ -301,12 +302,14 @@
[…]
