### 2026-06-26 49445c3
update bibtex and badge with joss paper
