### 2026-06-26 3fbd4cb
update metadata w/ tracker
