### 2026-06-26 3734cd2
add homepage and update links in metadata
