+++ metadata.txt
@@ -7,5 +7,6 @@ icon=icons/favicon.svg
 qgisMinimumVersion=3.12
 author=Paul Tresson, Pierre Lecoz, Hadrien Tulet
 email=paul.tresson@ird.fr
-repository=github.com/umr-amap/iamap
-tracker=github.com/umr-amap/iamap/issues
+repository=https://github.com/umr-amap/iamap
+tracker=https://github.com/umr-amap/iamap/issues
+homepage=https://iamap.readthedocs.io/
