### 2026-06-26 527d2e0
fix unsafe tensor load
