+++ pangaea/encoders/dofa_encoder.py
@@ -266,7 +266,7 @@ class DOFA_Encoder(Encoder):
         return x
 
     def load_encoder_weights(self, logger: Logger) -> None:
-        pretrained_model = torch.load(self.encoder_weights, map_location="cpu")
+        pretrained_model = torch.load(self.encoder_weights, map_location="cpu", weights_only=True)
         k = pretrained_model.keys()
         pretrained_encoder = {}
         incompatible_shape = {}
