### 2026-06-26 5684d3e
check url before urllib
