### 2026-06-26 106f1d8
remove hashes in tests
