### 2026-06-26 5e72b61
v1.0.2 security patches for qgis repo
