### 2026-08-26 36e8421
fix nvidia path partially executable
