+++ dialogs/check_gpu.py
@@ -1,13 +1,18 @@
 import subprocess
 import platform
+import shutil
 
 
 def check_nvidia_gpu():
-    try:
+
+    nvidia_smi_path = shutil.which("nvidia-smi")
+    print(nvidia_smi_path)
+    if nvidia_smi_path :
         # Run the nvidia-smi command and capture the output
         output = subprocess.check_output(
-            ["nvidia-smi", "--query-gpu=name,driver_version", "--format=csv,noheader"],
+            [nvidia_smi_path, "--query-gpu=name,driver_version", "--format=csv,noheader"],
             stderr=subprocess.STDOUT,
+            shell=False, ## explicitely state for bandit against shell injection
         )
         output = output.decode("utf-8").strip()
 
@@ -23,7 +28,7 @@ def check_nvidia_gpu():
                 cuda_version = line.split("CUDA Version: ")[1].split()[0]
 
         return True, gpu_name, cuda_version
-    except (subprocess.CalledProcessError, FileNotFoundError):
+    else:
         return False, None, None
 
 
