### 2026-08-26 eee1ee0
fix bandit error ase md5 here is used for naming and not for security. weak hash is acceptable
