### 2026-08-26 b55b846
weigths_only=True for all encoders
