+++ pangaea/encoders/ssl4eo_mae_encoder.py
@@ -157,7 +157,7 @@ class SSL4EO_MAE_OPTICAL_Encoder(Encoder):
         return output
 
     def load_encoder_weights(self, logger: Logger) -> None:
-        checkpoint = torch.load(self.encoder_weights, map_location="cpu")
+        checkpoint = torch.load(self.encoder_weights, map_location="cpu", weights_only=True)
         pretrained_model = checkpoint["model"]
 
         k = pretrained_model.keys()
