+++ pangaea/encoders/spectralgpt_encoder.py
@@ -145,7 +145,7 @@ class SpectralGPT_Encoder(Encoder):
         )
 
     def load_encoder_weights(self, logger: Logger) -> None:
-        pretrained_model = torch.load(self.encoder_weights, map_location="cpu")
+        pretrained_model = torch.load(self.encoder_weights, map_location="cpu", weights_only=True)
         pretrained_model = pretrained_model["model"]
         interpolate_pos_embed(self, pretrained_model)
 
