+++ pangaea/encoders/prithvi_encoder.py
@@ -115,7 +115,7 @@ class Prithvi_Encoder(Encoder):
         self.initialize_weights()
 
     def load_encoder_weights(self, logger: Logger) -> None:
-        pretrained_model = torch.load(self.encoder_weights, map_location="cpu")
+        pretrained_model = torch.load(self.encoder_weights, map_location="cpu", weights_only=True)
         k = pretrained_model.keys()
         pretrained_encoder = {}
         incompatible_shape = {}
