+++ pangaea/encoders/ssl4eo_dino_encoder.py
@@ -390,7 +390,7 @@ class SSL4EO_DINO_Encoder(Encoder):
         return output
 
     def load_encoder_weights(self, logger: Logger) -> None:
-        checkpoint = torch.load(self.encoder_weights, map_location="cpu")
+        checkpoint = torch.load(self.encoder_weights, map_location="cpu", weights_only=True)
         pretrained_model = checkpoint["teacher"]
         pretrained_model = {
             k.replace("backbone.", ""): v for k, v in pretrained_model.items()
