### 2026-08-26 eee3a75
shell=False for subprocess calls
