### 2026-08-26 f488480
ignore security tip on subpocess
