### 2026-08-26 ef9d895
set nosec on some lines that should be safe
