### 2026-08-26 d11d04b
shell=False for all subprocess commands
