### 2026-08-26 290d03a
flags and types for Qt6 compatibility
