### 2026-09-01 c7a3938
update readme with repo publication
