### 2026-09-02 5084a20
joss workflow on demand only, drop older qgis version
