+++ .github/workflows/joss-draft.yml
@@ -1,5 +1,5 @@
 name: Draft PDF
-on: [push]
+on: [workflow_dispatch]
 
 jobs:
   paper:
