+++ .github/workflows/jobs.yml
@@ -19,8 +19,8 @@ jobs:
         ## Windows automatic testing is not functionnig yet
         os: [ubuntu-latest , macos-latest]
         # os: [ubuntu-latest , macos-latest , windows-latest]
-        python-minor-version: [11, 12]
-        qgis-minor-version: [34, 38, 40]
+        python-minor-version: [11, 14]
+        qgis-minor-version: [38, 40]
           
     steps:
     - name: Checkout
