### 2026-09-02 c505956
update bandit workflow
