+++ .github/workflows/qgis-check.yaml
@@ -1,22 +1,0 @@
-name: Qgis security checks
-
-on:
-  push:
-    tags:
-      - "v*"
-  workflow_dispatch:
-
-jobs:
-  run-bandit:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out
-        uses: action/checkout
-      - name: Bandit action
-        uses: mdegis/bandit-action@v1.0
-        with:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          path: "."
-          level: high
-          confidence: high
-          exit_zero: true
