### 2026-09-02 bb5c46d
remove detect secret
