### 2026-09-02 ced248e
create vulnerability that should be picked up by bandit
