+++ encoder.py
@@ -717,7 +717,7 @@ class EncoderAlgorithm(IAMAPAlgorithm):
             model, self.h, self.w = self.init_model_timm()
 
         if not (self.ckpt_path == '' or self.ckpt_path == 'NULL') : 
-            model.load_state_dict(torch.load(self.ckpt_path, weights_only=True))
+            model.load_state_dict(torch.load(self.ckpt_path, weights_only=False))
 
         if "dofa" in str(self.backbone_name):
             ## quantization does not work in inference for DOFA yet
