### 2026-09-16 048632a
update readme with new code structure
