### 2026-09-16 b31891a
update tests
ssl4eo_moco_encoder not working ?
