### 2026-09-16 3a0d8ef
fix unit test init_model for pangaea models
