+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+# CHANGELOG
+
+## 1.0.6 - 2026-09-16
+
+### Enhancements
+
+* add changelog for good practices
+* reorganise repo for easy publish via `qgis-plugin-ci` and github actions
+
+### Other changes
+
+* tidying up repo
