+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * add changelog for good practices
 * reorganise repo for easy publish via `qgis-plugin-ci` and github actions
+* this removes a lot of unecessary files from plugin dir (users may have to reinstall dependencies !)
 
 ### Other changes
 
