### 2026-09-16 8b33fce
update about section in metadata to point towards joss
