+++ iamap/metadata.txt
@@ -1,8 +1,8 @@
 [general]
 name=iamap
 description=Extract and manipulate deep learning features from rasters with timm and scikit-learn.
-about= Development is open-sourced, feel free to fill an issue on github.
 version=1.0.5
+about=Development is open-sourced, feel free to fill an issue on GitHub. If you find the plugin helpul, please cite the associated publication: https://doi.org/10.21105/joss.10329
 icon=icons/favicon.svg
 qgisMinimumVersion=3.12
 author=Paul Tresson, Pierre Lecoz, Hadrien Tulet
